backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
openai>=1.0.0
python-dotenv>=1.0
numpy>=1.22
